docs and some refactoring
