+++ pkg/gatherers/conditional/conditional_gatherer.go
@@ -1,12 +1,10 @@
-// conditional package provides conditional gatherer which runs gatherings based on the rules and only if the provided
-// conditions are satisfied. Right now the rules are in the code (see `gatheringRules` at line 32), but later
-// they can be fetched from outside, checked that they make sense (we want to check the parameters, for example if
-// a rule tells to collect logs of a namespace on firing alert, we want to check that the namespace is created
-// by openshift and not by a user). Conditional gathering isn't considered prioritized, so we run it every 6 hours.
+// Package conditional provides conditional gatherer which runs gatherings based on the rules and only if the provided
+// conditions are satisfied. The rules are fetched from Insights Operator Gathering Conditions Service
+// https://github.com/RedHatInsights/insights-operator-gathering-conditions-service . The rules are validated to
+// check that they make sense (for example we don't allow collecting logs from non openshift namespaces).
 //
-// To add a new condition, follow the next steps:
-// 1. Add structures to conditions.go
-// 2. Change areAllConditionsSatisfied function in conditional_gatherer.go
+// To add a new condition, follow the steps described in conditions.go file.
+// To add a new gathering function, follow the steps described in gathering_functions.go file.
 package conditional
 
 import (
@@ -16,7 +14,6 @@ import (
 	"sort"
 	"strings"
 
-	"github.com/blang/semver/v4"
 	configv1client "github.com/openshift/client-go/config/clientset/versioned/typed/config/v1"
 	"github.com/prometheus/common/expfmt"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
@@ -29,16 +26,6 @@ import (
 	"github.com/openshift/insights-operator/pkg/record"
 	"github.com/openshift/insights-operator/pkg/utils"
 )
-
-// gatheringFunctionBuilders lists all the gatherers which can be run on some condition. Gatherers can have parameters,
-// like namespace or number of log lines to fetch, see the docs of the functions.
-var gatheringFunctionBuilders = map[GatheringFunctionName]GathererFunctionBuilderPtr{
-	GatherLogsOfNamespace:         (*Gatherer).BuildGatherLogsOfNamespace,
-	GatherImageStreamsOfNamespace: (*Gatherer).BuildGatherImageStreamsOfNamespace,
-	GatherAPIRequestCounts:        (*Gatherer).BuildGatherAPIRequestCounts,
-	GatherContainersLogs:          (*Gatherer).BuildGatherContainersLogs,
-	GatherPodDefinition:           (*Gatherer).BuildGatherPodDefinition,
-}
 
 // Gatherer implements the conditional gatherer
 type Gatherer struct {
@@ -193,35 +180,6 @@ func (g *Gatherer) getGatheringRulesJSON(ctx context.Context) (string, error) {
 	return string(rulesBytes), err
 }
 
-// areAllConditionsSatisfied returns true if all the conditions are satisfied, for example if the condition is
-// to check if a metric is firing, it will look at that metric and return the result according to that
-func (g *Gatherer) areAllConditionsSatisfied(conditions []ConditionWithParams) (bool, error) {
-	for _, condition := range conditions {
-		switch condition.Type {
-		case AlertIsFiring:
-			if condition.Alert == nil {
-				return false, fmt.Errorf("alert field should not be nil")
-			}
-
-			if firing, err := g.isAlertFiring(condition.Alert.Name); !firing || err != nil {
-				return false, err
-			}
-		case ClusterVersionMatches:
-			if condition.ClusterVersionMatches == nil {
-				return false, fmt.Errorf("cluster_version_matches field should not be nil")
-			}
-
-			if doesMatch, err := g.doesClusterVersionMatch(condition.ClusterVersionMatches.Version); !doesMatch || err != nil {
-				return false, err
-			}
-		default:
-			return false, fmt.Errorf("unknown condition type: %v", condition.Type)
-		}
-	}
-
-	return true, nil
-}
-
 // updateCache updates alerts and version caches
 func (g *Gatherer) updateCache(ctx context.Context) {
 	if g.metricsGatherKubeConfig == nil {
@@ -309,39 +267,7 @@ func (g *Gatherer) updateVersionCache(ctx context.Context, configClient configv1
 	return nil
 }
 
-// isAlertFiring using the cache it returns true if the alert is firing
-func (g *Gatherer) isAlertFiring(alertName string) (bool, error) {
-	if g.firingAlerts == nil {
-		return false, fmt.Errorf("alerts cache is missing")
-	}
-
-	_, alertIsFiring := g.firingAlerts[alertName]
-	return alertIsFiring, nil
-}
-
-func (g *Gatherer) doesClusterVersionMatch(expectedVersionExpression string) (bool, error) {
-	if len(g.clusterVersion) == 0 {
-		return false, fmt.Errorf("cluster version is missing")
-	}
-
-	clusterVersion, err := semver.Parse(g.clusterVersion)
-	if err != nil {
-		return false, err
-	}
-
-	expectedRange, err := semver.ParseRange(expectedVersionExpression)
-	if err != nil {
-		return false, err
-	}
-
-	// ignore everything after the first three numbers
-	clusterVersion.Pre = nil
-	clusterVersion.Build = nil
-
-	return expectedRange(clusterVersion), nil
-}
-
-// createGatheringClosures produces gathering closures from the rules
+// createGatheringClosures produces gathering closures
 func (g *Gatherer) createGatheringClosures(
 	gatheringFunctions map[GatheringFunctionName]interface{},
 ) (map[string]gatherers.GatheringClosure, []error) {
